requirements.txt
customtkinter
packaging
pillow
pyaudio
moviepy
requests
torch==1.12.0
tqdm
modelscope
numpy
openai-whisper
basicsr==1.4.2
beautifulsoup4==4.12.2
opencv-python==4.9
facenet_pytorch
g2p-en==2.10
gfpgan==1.3.8
jieba==0.42.1